% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimJTC_.R
\name{optimTP.GA}
\alias{optimTP.GA}
\title{function optimTP.GA}
\usage{
optimTP.GA(
  ncores,
  formula,
  miscov,
  auxvar,
  family,
  n,
  data,
  beta,
  p_gz,
  disp = NULL,
  ga.popsize,
  ga.propelit,
  ga.proptourney,
  ga.ngen,
  ga.mutrate,
  ga.initpop = NULL,
  optimMeasure,
  K.idx = NULL,
  seed = 1,
  verbose = 0
)
}
\arguments{
\item{ncores}{ncores1}

\item{formula}{the formula}

\item{miscov}{miscov1}

\item{auxvar}{auxvar1}

\item{family}{family1}

\item{n}{n1}

\item{data}{the data gggg}

\item{beta}{beta1}

\item{p_gz}{p_gz1}

\item{disp}{disp1}

\item{ga.popsize}{ga.popsize1}

\item{ga.propelit}{ga.propelit1}

\item{ga.proptourney}{ga.proptourney1}

\item{ga.ngen}{ga.ngen1}

\item{ga.mutrate}{ga.mutrate1}

\item{ga.initpop}{ga.initpop1}

\item{optimMeasure}{optimMeasure1}

\item{K.idx}{K.idx1}

\item{seed}{seed1}

\item{verbose}{verbose1}
}
\description{
function optimTP.GA
}
\details{
details here
}
\examples{

print(1)

}
