% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimJTC_.R
\name{optimTP.LM}
\alias{optimTP.LM}
\title{function optimTP.LM}
\usage{
optimTP.LM(
  formula,
  miscov,
  auxvar,
  strata,
  family,
  n,
  data,
  beta,
  p_gz,
  disp = NULL,
  optimMeasure,
  K.idx = NULL,
  min.nk = NULL,
  logical.sub = NULL
)
}
\arguments{
\item{formula}{formula1name}

\item{miscov}{xxxxxxxxcode}

\item{auxvar}{auxvar1}

\item{strata}{strata1}

\item{family}{family1}

\item{n}{n1}

\item{data}{data1}

\item{beta}{beta1}

\item{p_gz}{p_gz1}

\item{disp}{disp1}

\item{optimMeasure}{optimMeasure1}

\item{K.idx}{K.idx1}

\item{min.nk}{min.nk1}

\item{logical.sub}{logical.sub1}
}
\description{
function optimTP.LM
}
\details{
details at here
}
\examples{

print(1)

}
