% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoPhaseSPMLglm_.R
\name{twoPhaseSPML}
\alias{twoPhaseSPML}
\title{Performs inference on two-phase studies data via semiparametric maximum likelihood.}
\usage{
twoPhaseSPML(
  formula,
  miscov,
  auxvar,
  family = gaussian,
  data0,
  data1,
  start.values = NULL,
  verbose = FALSE
)
}
\arguments{
\item{formula}{regression formula, note that if it does not contain the missing-by-design variable, miscov, it will return results under the null hypothesis. Hypothesis testing corresponds to the score statistic. Otherwise, estimates and hypothesis testing ocurr under the alternative hypothesis leading to Wald statistics. All the elements in formula except miscov must be present in data0 and data1.}

\item{miscov}{right hand side formula with the missing-by-design covariate(s), i.e. the potential causal locus (loci). Must be present in data1 but absent in data0.}

\item{auxvar}{right hand side formula with the auxiliary variable(s), i.e. the GWAS SNP from phase 1. Must be present in data0 and data1.}

\item{family}{member of the exponential family (see \code{\link{family}}, 'quasi' models not available). Default gaussian().}

\item{data0}{a dataframe with the complement of the phase 2 data. Must contain the unique elements in formula and auxvar but NOT miscov.}

\item{data1}{a dataframe with the phase 2 data. Must contain the unique elements in formula, auxvar, and miscov.}

\item{start.values}{a named list with initial values for the regression parameters and joint distribution between miscov and auxvar (only one can be specified). Defaults to NULL}

\item{verbose}{verbose output? logical, defaults to FALSE.}
}
\value{
A list of objects
}
\description{
Performs inference on two-phase studies data via semiparametric maximum likelihood.
}
\details{
these are some additional details
}
\examples{

data = DataGeneration_TPD()
set.seed(1)
R = rep(0, nrow(data)); R[sample(nrow(data),500)] <- 1 # random phase 2 subsample of 500.
data0 = data[R==0,c('Y','Z')]
data1 = data[R==1,c('Y','Z','G1')]
res_Ho = twoPhaseSPML(formula =  Y ~ Z,
miscov = ~ G1,
auxvar = ~ Z,
data0 = data0, data1 = data1)
res_Ha = twoPhaseSPML(formula =  Y ~ Z + G1,
miscov = ~ G1,
auxvar = ~ Z,
data0 = data0, data1 = data1)
}
